requirements.txt
asgiref==3.2.3
attrs==19.3.0
coverage==5.0.3
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
importlib-metadata==1.5.0
more-itertools==8.2.0
mysqlclient==1.4.6
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wcwidth==0.1.8
zipp==3.1.0
